debug uthread.c
cast address to uint64

+++ Thread/user/uthread.c
@@ -81,7 +81,7 @@ thread_schedule(void)
      * Invoke thread_switch to switch from t to next_thread:
      * thread_switch(??, ??);
      */
-    thread_switch(&t->context,&current_thread->context);
+    thread_switch((uint64)&(t->context),(uint64)&(current_thread->context));
     t->state = RUNNABLE;
   } else
     next_thread = 0;
